Use new binded save device functions in python
