+++ parsec/core/cli/list_devices.py
@@ -29,7 +29,7 @@ def list_devices(
         config_dir = Path(config_dir) if config_dir else get_default_config_dir(os.environ)
         devices = list_available_devices(config_dir)
         num_devices_display = click.style(str(len(devices)), fg="green")
-        config_dir_display = click.style(str(config_dir), fg="yellow")
+        config_dir_display = click.style(config_dir, fg="yellow")
         click.echo(f"Found {num_devices_display} device(s) in {config_dir_display}:")
         click.echo(format_available_devices(devices))
