+++ parsec/core/gui/claim_device_widget.py
@@ -10,7 +10,7 @@ from PyQt5.QtGui import QFont
 from PyQt5.QtWidgets import QWidget
 from structlog import get_logger
 
-from parsec._parsec import SASCode
+from parsec._parsec import SASCode, save_device_with_password_in_config
 from parsec.api.protocol import DeviceLabel
 from parsec.core import CoreConfig
 from parsec.core.backend_connection import (
@@ -35,7 +35,6 @@ from parsec.core.local_device import (
     LocalDeviceCryptoError,
     LocalDeviceError,
     LocalDeviceNotFoundError,
-    save_device_with_password_in_config,
     save_device_with_smartcard_in_config,
 )
 from parsec.core.types import BackendInvitationAddr, LocalDevice
