+++ tests/core/gui/test_main_window.py
@@ -6,7 +6,12 @@ from unittest.mock import patch
 import pytest
 from PyQt5 import QtCore, QtGui, QtWidgets
 
-from parsec._parsec import DateTime, InvitationType
+from parsec._parsec import (
+    DateTime,
+    InvitationType,
+    save_device_with_password,
+    save_device_with_password_in_config,
+)
 from parsec.api.data import EntryName
 from parsec.api.protocol import OrganizationID, UserProfile
 from parsec.core.fs.workspacefs import WorkspaceFSTimestamped
@@ -14,12 +19,7 @@ from parsec.core.gui import desktop
 from parsec.core.gui.lang import translate
 from parsec.core.gui.login_widget import LoginPasswordInputWidget
 from parsec.core.gui.workspace_roles import get_role_translation as _
-from parsec.core.local_device import (
-    AvailableDevice,
-    DeviceFileType,
-    save_device_with_password,
-    save_device_with_password_in_config,
-)
+from parsec.core.local_device import AvailableDevice, DeviceFileType
 from parsec.core.types import (
     BackendInvitationAddr,
     BackendOrganizationAddr,
