+++ tests/core/gui/test_recovery.py
@@ -6,6 +6,7 @@ from pathlib import Path
 import pytest
 from PyQt5 import QtCore
 
+from parsec._parsec import save_device_with_password_in_config
 from parsec.api.protocol import DeviceLabel
 from parsec.core.gui.authentication_choice_widget import AuthenticationChoiceWidget
 from parsec.core.gui.device_recovery_export_widget import (
@@ -14,11 +15,7 @@ from parsec.core.gui.device_recovery_export_widget import (
 )
 from parsec.core.gui.device_recovery_import_widget import DeviceRecoveryImportPage1Widget
 from parsec.core.gui.lang import translate
-from parsec.core.local_device import (
-    get_recovery_device_file_name,
-    save_device_with_password_in_config,
-    save_recovery_device,
-)
+from parsec.core.local_device import get_recovery_device_file_name, save_recovery_device
 from parsec.core.recovery import generate_recovery_device
